Refactor: (`vertix-gui`) - Merge `UIServiceAdapter` with `UIService`
