+++ packages/vertix-bot/src/ui-v2/_general/admin/admin-adapter-exu-base.ts
@@ -57,7 +57,7 @@ export class AdminAdapterExuBase<
                     `🔐 Bot missing permissions" - "${ missingPermissions.join( ", " ) }" (${ interaction.guild.name }) (${ interaction.guild?.memberCount })`
                 );
 
-                await this.uiAdapterService.get( "VertixGUI/InternalAdapters/MissingPermissionsAdapter" )?.ephemeral( interaction, {
+                await this.uiService.get( "VertixGUI/InternalAdapters/MissingPermissionsAdapter" )?.ephemeral( interaction, {
                     missingPermissions,
                     omitterDisplayName: interaction.guild.client.user.username,
                 } );
