+++ packages/vertix-bot/src/ui-v2/setup-edit/setup-edit-adapter.ts
@@ -208,7 +208,7 @@ export class SetupEditAdapter extends AdminAdapterExuBase<VoiceChannel, Interact
     }
 
     protected async regenerate( interaction: MessageComponentInteraction<"cached"> ): Promise<void> {
-        this.uiAdapterService.get( "VertixBot/UI-V2/SetupAdapter" )?.editReply( interaction );
+        this.uiService.get( "VertixBot/UI-V2/SetupAdapter" )?.editReply( interaction );
     }
 
     private async onSetupMasterEditButtonClicked( interaction: UIDefaultButtonChannelTextInteraction ) {
@@ -336,7 +336,7 @@ export class SetupEditAdapter extends AdminAdapterExuBase<VoiceChannel, Interact
             case "VertixBot/UI-V2/SetupEditMaster":
                 this.deleteArgs( interaction );
 
-                this.uiAdapterService.get( "VertixBot/UI-V2/SetupAdapter" )?.editReply( interaction );
+                this.uiService.get( "VertixBot/UI-V2/SetupAdapter" )?.editReply( interaction );
                 break;
         }
