+++ packages/vertix-gui/src/internal-adapters/invalid-channel-type-adapter.ts
@@ -5,9 +5,8 @@ import { UIAdapterBase } from "@vertix.gg/gui/src/bases/ui-adapter-base";
 import type { UIArgs } from "@vertix.gg/gui/src/bases/ui-definitions";
 import type { UIDefaultButtonChannelVoiceInteraction } from "@vertix.gg/gui/src/bases/ui-interaction-interfaces";
 
-import type { UIAdapterService } from "@vertix.gg/gui/src/ui-adapter-service";
-
 import type { BaseGuildTextChannel } from "discord.js";
+import type { UIService } from "@vertix.gg/gui/src/ui-service";
 
 export class InvalidChannelTypeAdapter extends UIAdapterBase<BaseGuildTextChannel, UIDefaultButtonChannelVoiceInteraction> {
     public static getName() {
@@ -15,7 +14,7 @@ export class InvalidChannelTypeAdapter extends UIAdapterBase<BaseGuildTextChanne
     }
 
     public static getComponent() {
-        const Component = ServiceLocator.$.get<UIAdapterService>( "VertixGUI/UIAdapterService" ).$$
+        const Component = ServiceLocator.$.get<UIService>( "VertixGUI/UIService" ).$$
             .getSystemComponents().InvalidChannelTypeComponent;
 
         if ( ! Component ) {
