+++ packages/vertix-gui/src/internal-adapters/missing-permissions-adapter.ts
@@ -6,8 +6,7 @@ import type { BaseGuildTextChannel } from "discord.js";
 
 import type { UIArgs } from "@vertix.gg/gui/src/bases/ui-definitions";
 import type { UIDefaultButtonChannelVoiceInteraction } from "@vertix.gg/gui/src/bases/ui-interaction-interfaces";
-
-import type UIAdapterService from "@vertix.gg/gui/src/ui-adapter-service";
+import type { UIService } from "@vertix.gg/gui/src/ui-service";
 
 export class MissingPermissionsAdapter extends UIAdapterBase<BaseGuildTextChannel, UIDefaultButtonChannelVoiceInteraction> {
     public static getName() {
@@ -15,7 +14,7 @@ export class MissingPermissionsAdapter extends UIAdapterBase<BaseGuildTextChanne
     }
 
     public static getComponent() {
-        const Component = ServiceLocator.$.get<UIAdapterService>( "VertixGUI/UIAdapterService" ).$$
+        const Component = ServiceLocator.$.get<UIService>( "VertixGUI/UIService" ).$$
             .getSystemComponents().MissingPermissionsComponent;
 
         if ( ! Component ) {
