+++ packages/vertix-gui/test/_base/ui-template-base.spec.ts
@@ -1,4 +1,4 @@
-import { ServiceLocatorMock } from "@vertix.gg/test-utils/src/__mock__/service-locator-mock";
+import { TestWithServiceLocatorMock } from "@vertix.gg/test-utils/src/test-with-service-locator-mock";
 
 import { UITemplateBase } from "@vertix.gg/gui/src/bases/ui-template-base";
 
@@ -20,20 +20,9 @@ describe( "VertixGUI/UITemplateBase", () => {
     let templateBase: MyTemplate;
 
     beforeEach( async () => {
-        // Mock original ServiceLocator.
-        ServiceLocatorMock.mockOrigin();
-
-        ServiceLocatorMock.$.register( ( await import( "@vertix.gg/test-utils/src/__mock__/ui-service-mock" ) ).UIServiceMock );
-
-        // Await for all services to be registered.
-        await ServiceLocatorMock.$.waitForAll();
+        await TestWithServiceLocatorMock.withUIServiceMock();
 
         templateBase = new MyTemplate;
-    } );
-
-    afterEach( () => {
-        // Reset ServiceLocator.
-        ServiceLocatorMock.reset();
     } );
 
     test( "composeTemplate() :: should handle missing selectOptions gracefully", () => {
