+++ packages/vertix-test-utils/src/__mock__/ui-service-mock.ts
@@ -1,6 +1,6 @@
 import Mocker from "ts-mockito";
 
-import UIService from "@vertix.gg/gui/src/ui-service";
+import { UIService } from "@vertix.gg/gui/src/ui-service";
 
 import type { Client } from "discord.js";
